backend/requirements.txt
# ──── FastAPI Core ────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ──── Data Validation ────
pydantic>=2.7.0

# ──── Environment configuration ────
python-dotenv>=1.0.0

# ──── Async HTTP & Testing ────
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
anyio>=4.4.0

# ──── ML Inference (structure only — weights loaded separately) ────
torch>=2.0.0
ultralytics>=8.0.0
transformers>=4.30.0
Pillow>=10.0.0
numpy>=1.24.0

# ──── Database ────
sqlalchemy[asyncio]>=2.0
psycopg[binary]>=3.1
geoalchemy2>=0.14
alembic>=1.13

# ──── Settings ────
pydantic-settings>=2.0

# ──── Image hashing ────
imagehash>=4.3

# ──── Testing extras ────
hypothesis>=6.100

